move CommandCreateFP to a separate file
